Made changes to entity defns, added repositories for group and groupMembership and impolemented groupService

+++ src/main/java/com/zealmobile/studygroup/core/entities/Group.java
@@ -30,7 +30,7 @@ public class Group {
 	private String name;
 
 	@Column(name="owner_id")
-	private int ownerId;
+	private Long ownerId;
 
 	@Column(name="created_date")
 	private Date createDate;
